added new premodel 3d with attn

+++ lit_models/Vesuvius_Lit_Model.py
@@ -297,7 +297,7 @@ class TverskyLoss(nn.Module):
         # comment out if your model contains a sigmoid or equivalent activation layer
 
         inputs = F.sigmoid(inputs)
-        # inputs = (inputs >= 0.4)
+        # inputs = (inputs >= 0.4).float()
 
         # flatten label and prediction tensors
         inputs = inputs.view(-1)
@@ -318,7 +318,7 @@ def dice_coef_torch(preds, targets, beta=0.5, smooth=1e-1, threshold = 0.4):
 
     #comment out if your model contains a sigmoid or equivalent activation layer
     preds = torch.sigmoid(preds)
-    preds = (preds >= threshold)
+    preds = (preds >= threshold).float()
 
     # flatten label and prediction tensors
     preds = preds.view(-1).float()
